Refactor gogit test into its own file
